Use "resolve" for polymorphic hierarchy paths
In the past, we couldn't use `polymorphic_path` in many places. For
instance, `polymorphic_path(budget, investment)` would return
`budget_budget_investment_path`, while in our routes we had defined
`budget_investment_path`.

With the `resolve` method, introduced in Rails 5.1, we can use symbols
to define we want it to use `investment` instead of `budget_investment`.
It also works with nested resources, so now we can write
`polymorphic_path(investment)`.

This makes the code for `resource_hierarchy_for` almost impossible to
understand. I reached this result after having a look at the internals
of the `resolve` method in order to get its results and then remove the
symbols we include.

Note using this method will not make admin routes compatible with
`polymorphic_path`. Quoting from the Rails documentation:

> This custom behavior only applies to simple polymorphic URLs where a
> single model instance is passed and not more complicated forms, e.g:
> [example showing admin routes won't work]

Also note that now the `admin_polymorphic_path` method will not work for
every model due to inconsistencies in our admin routes. For instance, we
define `groups` and `budget_investments`; we should either use the
`budget_` prefix in all places or remove it everywhere. Right now the
code only works for items with the prefix; it isn't a big deal because
we never call it with an item without the prefix.

The following code would only work for items without the prefix.

```
def admin_polymorphic_path(resource, options = {})
  resolve = resolve_for(resource)

  if resolve
    if resolve.last.is_a?(Hash)
      polymorphic_path([:admin, *resolve[0..-2]], options.merge(resolve.last))
    else
      polymorphic_path([:admin, resolve], options)
    end
  else
    polymorphic_path([:admin, resource], options)
  end
end

private

  def resolve_for(resource)
    polymorphic_mapping(resource)&.send(:eval_block, self, resource, {})
  end
```